Moved shared DLQ header behavior to `RetryStrategy`
